refactor: Move components representing specific visuals to graphs folder
